fix(bin-designer): close ghost glyph outlines and retry a failed font load
Three review findings, each of which fails quietly:

- The ghost overlay walked a contour's points without the closing run back to
  the start, so every letter was drawn one edge short. A glyph contour ends
  with a `Z` that carries no coordinates.
- A font whose fetch failed stayed recorded as in flight on both the worker
  and the main thread, so one flaky request cost that face for the rest of
  the session with no retry and, on the worker, no warning.
- Two suites read font files by a path relative to the process CWD.
